requirements.txt
beautifulsoup4
facebook-scraper
insta-scraper
linkedin-scraper
mysql-connector-python
requests
selenium
snscrape
matplotlib
nltk
pandas
numpy
seaborn
scipy
scikit-learn
textblob
python-dotenv